+++ tests/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+using Xunit;
+
+// Disable parallel test execution for the entire assembly
+[assembly: CollectionBehavior(DisableTestParallelization = true)]
