+++ tests/tests.csproj
@@ -5,6 +5,9 @@
     <PlatformTarget Condition="$(PlatformTarget) == ''">x64</PlatformTarget>
     <IsPackable>false</IsPackable>
     <XunitConsole>$(NuGetPackageRoot)xunit.runner.console/2.4.2/tools/net472/xunit.console.exe</XunitConsole>
+    <!-- Ensure xUnit doesn't run tests in parallel -->
+    <XunitParallelizeAssembly>false</XunitParallelizeAssembly>
+    <XunitParallelizeTestCollections>false</XunitParallelizeTestCollections>
   </PropertyGroup>
 
   <PropertyGroup Condition="$(PlatformTarget) == 'x64'">
